+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Add basic support for DOJ 2026-01-30 files
 
 ### 1.4.1
 * Use `inkscape` instead of `cairosvg` for .png emailers table generation